Separate Indexer.Address.TokenBalances from Indexer.Balances
TokenBalances need to extract 3 fields, while Balances only need 2, so
they shouldn't be mixed together in the same module since the expected
params in the set differs.
