+++ apps/indexer/test/indexer/balances_test.exs
@@ -107,36 +107,6 @@ defmodule Indexer.BalancesTest do
       assert %{address_hash: address_hash, block_number: block_number}
     end
 
-    test "with token transfer extract from_address, to_address, and token_contract_address_hash" do
-      block_number = 1
-
-      from_address_hash =
-        Factory.address_hash()
-        |> to_string()
-
-      to_address_hash =
-        Factory.address_hash()
-        |> to_string()
-
-      token_contract_address_hash =
-        Factory.address_hash()
-        |> to_string()
-
-      token_transfer_params = %{
-        block_number: block_number,
-        from_address_hash: from_address_hash,
-        to_address_hash: to_address_hash,
-        token_contract_address_hash: token_contract_address_hash
-      }
-
-      params_set = Balances.params_set(%{token_transfers_params: [token_transfer_params]})
-
-      assert MapSet.size(params_set) == 3
-      assert %{address_hash: from_address_hash, block_number: block_number}
-      assert %{address_hash: to_address_hash, block_number: block_number}
-      assert %{address_hash: token_contract_address_hash, block_number: block_number}
-    end
-
     test "with transaction without to_address_hash extracts from_address_hash" do
       block_number = 1
